clean up warnings and get android compiling git st
