+++ android/src/main/java/expo/modules/xmtpreactnativesdk/wrappers/GroupWrapper.kt
@@ -2,8 +2,8 @@ package expo.modules.xmtpreactnativesdk.wrappers
 
 import com.google.gson.GsonBuilder
 import com.google.gson.JsonParser
-import expo.modules.xmtpreactnativesdk.wrappers.ConsentWrapper.Companion.consentStateToString
 import org.xmtp.android.library.Client
+import org.xmtp.android.library.ConsentState
 import org.xmtp.android.library.Group
 
 class GroupWrapper {
@@ -46,6 +46,14 @@ class GroupWrapper {
             val obj = encodeToObj(client, group, groupParams)
             return gson.toJson(obj)
         }
+    }
+}
+
+fun consentStateToString(state: ConsentState): String {
+    return when (state) {
+        ConsentState.ALLOWED -> "allowed"
+        ConsentState.DENIED -> "denied"
+        ConsentState.UNKNOWN -> "unknown"
     }
 }
