+++ android/src/main/java/expo/modules/xmtpreactnativesdk/wrappers/MemberWrapper.kt
@@ -1,12 +1,13 @@
 package expo.modules.xmtpreactnativesdk.wrappers
 
 import com.google.gson.GsonBuilder
+import org.xmtp.android.library.ConsentState
 import org.xmtp.android.library.libxmtp.Member
 import org.xmtp.android.library.libxmtp.PermissionLevel
 
 class MemberWrapper {
     companion object {
-        fun encodeToObj(member: Member): Map<String, Any> {
+        private fun encodeToObj(member: Member): Map<String, Any> {
             val permissionString = when (member.permissionLevel) {
                 PermissionLevel.MEMBER -> "member"
                 PermissionLevel.ADMIN -> "admin"
@@ -16,7 +17,7 @@ class MemberWrapper {
                 "inboxId" to member.inboxId,
                 "addresses" to member.addresses,
                 "permissionLevel" to permissionString,
-                "consentState" to ConsentWrapper.consentStateToString(member.consentState)
+                "consentState" to consentStateToString(member.consentState)
             )
         }
